Returning non-zero status in case of error in parameters
Knowing when something happened correctly and when an error occurred is important when using zopflipng in other scripts.

In special, if the help message is printed without passing `-h` or `--help`, it happened because of wrong or missing parameters, and should return a non-zero status. However, passing `-h` or `--help` will return 0.