Properly specify prerequisites in Makefile
Previously, prerequisites for some targets were implicit in nature,
meaning that whether the library targets actually attempted to link
with object files compiled with -fPIC depended on the leftover files
from previous build and the order in which targets were built.

This could lead to build failures for the libraries, and the standalone
utilities may have been able to be linked with object files compiled
with position independent code.

Furthermore, a target "all" was added to build all targets, and the
phony targets were properly set to "all" and "clean".

This fixes issue #68.