Use fstat(), not fseek(), to determine file size
Subclause 7.21.9.2 of the C standard stipulates that "a binary stream need not
meaningfully support fseek calls with a whence value of SEEK_END";
Additionally, footnote 268 of subclause 7.21.3 says: "Setting the file position
indicator to end-of-file, as with fseek(file, 0, SEEK_END), has undefined
behavior for a binary stream". On non-POSIX systems, using fseek() to set the
file position indicator to end-of-file is not guaranteed to work. If the
program tries to allocate memory based on the result of fseek() / ftell(), the
amount may be incorrect, leading to a potential vulnerability.

To address this issue, we can simply use the file size provided by fstat(),
instead of fseek() / ftell().

Both the diagnosis and the resolution are supported by recommendation FIO19-C
of the SEI CERT C Coding Standard.[1]

[1]: https://www.securecoding.cert.org/confluence/display/c/FIO19-C.+Do+not+use+fseek%28%29+and+ftell%28%29+to+compute+the+size+of+a+regular+file?src=contextnavpagetreemode