Fix excessive usage of malloc/free
Split ZopfliInitHash to:
- ZopfliAllocHash - allocate hash memory,
- ZopfliResetHash - reset hash values.

Allocate Hash outside of ZopfliLZ77Greedy and ZopfliLZ77OptimalRun that
pass it further to functions previously allocating them. Do the same for
costs malloc'd array.

Reason for this change:
- the size of malloc doesn't change,
- speed up Zopfli*,
- fix crash on certain devices**.

* speeds up Zopfli (especially on smaller blocks) by reducing amount of
sys time from ~7s to 0.1s on x64 Linux for ~5m compression time and from
~1m to 0.1s on ARMv7 Linux for 13m compression time.
** fixes a large amount of iterations crash on some ARM devices that due
to architecture or older kernel (not sure which) don't handle too
aggressive heap allocation and freeing.